Change 'playsong' implementation
Might be better overall for when I start hooking this thing up to Spotify as well. Still not sure how that's going to work, especially since I've got two sound libraries imported right now.
Also changed some various functions including removing Linux support for the openapp control. This is because I am unfamiliar with Linux and would rather not randomly type something in someone's console without their permission or my educated understanding of it.
Overall, this update is kind of a 'meh' to me.